+++ asyncdex/models/__init__.py
@@ -1,6 +1,8 @@
 from .author import Author
 from .chapter import Chapter, ChapterList
+from .custom_list import CustomList
 from .group import Group
 from .manga import Manga, MangaList
 from .tag import Tag
+from .title import TitleList
 from .user import User